call printf instead of ls
in this case, there's not need to send errors to stderr. we rely on
globing, which is less resource-hungry than a plain stat. if the file is
there and readable, printf happily shows it otherwise it doesn't. the
slash at the end is important as it means we're only interested in
directories (since you used the command ls -d before).